Convert plans page to table (still needs pagination and sorting)
